Refactor to dedicated client and gateway modules
