+++ rust/connlib/tunnel/src/gateway.rs
@@ -1,8 +1,16 @@
 use crate::device_channel::create_iface;
-use crate::{ControlSignal, Device, GatewayState, Tunnel, MAX_UDP_SIZE};
-use connlib_shared::messages::Interface as InterfaceConfig;
+use crate::{
+    ControlSignal, Device, Event, RoleState, Tunnel, ICE_GATHERING_TIMEOUT_SECONDS,
+    MAX_CONCURRENT_ICE_GATHERING, MAX_UDP_SIZE,
+};
+use connlib_shared::messages::{ClientId, Interface as InterfaceConfig};
 use connlib_shared::Callbacks;
+use futures::channel::mpsc::Receiver;
+use futures_bounded::{PushError, StreamMap};
 use std::sync::Arc;
+use std::task::{ready, Context, Poll};
+use std::time::Duration;
+use webrtc::ice_transport::ice_candidate::RTCIceCandidateInit;
 
 impl<C, CB> Tunnel<C, CB, GatewayState>
 where
@@ -54,5 +62,56 @@ where
             })
             .abort_handle(),
         );
+    }
+}
+
+/// [`Tunnel`](crate::Tunnel) state specific to gateways.
+pub struct GatewayState {
+    candidate_receivers: StreamMap<ClientId, RTCIceCandidateInit>,
+}
+
+impl GatewayState {
+    pub fn add_new_ice_receiver(&mut self, id: ClientId, receiver: Receiver<RTCIceCandidateInit>) {
+        match self.candidate_receivers.try_push(id, receiver) {
+            Ok(()) => {}
+            Err(PushError::BeyondCapacity(_)) => {
+                tracing::warn!("Too many active ICE candidate receivers at a time")
+            }
+            Err(PushError::Replaced(_)) => {
+                tracing::warn!(%id, "Replaced old ICE candidate receiver with new one")
+            }
+        }
+    }
+}
+
+impl Default for GatewayState {
+    fn default() -> Self {
+        Self {
+            candidate_receivers: StreamMap::new(
+                Duration::from_secs(ICE_GATHERING_TIMEOUT_SECONDS),
+                MAX_CONCURRENT_ICE_GATHERING,
+            ),
+        }
+    }
+}
+
+impl RoleState for GatewayState {
+    type Id = ClientId;
+
+    fn poll_next_event(&mut self, cx: &mut Context<'_>) -> Poll<Event<Self::Id>> {
+        loop {
+            match ready!(self.candidate_receivers.poll_next_unpin(cx)) {
+                (conn_id, Some(Ok(c))) => {
+                    return Poll::Ready(Event::SignalIceCandidate {
+                        conn_id,
+                        candidate: c,
+                    })
+                }
+                (id, Some(Err(e))) => {
+                    tracing::warn!(gateway_id = %id, "ICE gathering timed out: {e}")
+                }
+                (_, None) => {}
+            }
+        }
     }
 }
